Doctrine: Migration: added functionality for ignore column from entity by annotation
