+++ packages/framework/src/Component/Doctrine/RemoveEntityColumnSubscriber.php
@@ -1,0 +1,82 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shopsys\FrameworkBundle\Component\Doctrine;
+
+use Doctrine\Common\EventSubscriber;
+use Doctrine\ORM\Event\LoadClassMetadataEventArgs;
+use Doctrine\ORM\Events;
+use Doctrine\ORM\Mapping\ClassMetadata;
+use Shopsys\FrameworkBundle\Component\Doctrine\Attribute\RemoveColumn;
+
+class RemoveEntityColumnSubscriber implements EventSubscriber
+{
+    /**
+     * @return string[]
+     */
+    public function getSubscribedEvents(): array
+    {
+        return [
+            Events::loadClassMetadata,
+        ];
+    }
+
+    /**
+     * @param \Doctrine\ORM\Event\LoadClassMetadataEventArgs $eventArgs
+     */
+    public function loadClassMetadata(LoadClassMetadataEventArgs $eventArgs): void
+    {
+        $classMetadata = $eventArgs->getClassMetadata();
+
+        foreach ($classMetadata->getReflectionClass()->getProperties() as $property) {
+            $reflectionAttributes = $property->getAttributes(RemoveColumn::class);
+
+            if (count($reflectionAttributes) > 0) {
+                $this->removeColumnFromEntityMappings($property->getName(), $classMetadata);
+            }
+        }
+    }
+
+    /**
+     * @param string $propertyName
+     * @param \Doctrine\ORM\Mapping\ClassMetadata $classMetadata
+     */
+    protected function removeColumnFromEntityMappings(string $propertyName, ClassMetadata $classMetadata): void
+    {
+        $classMetadata->associationMappings = $this->removeMappingByKey($propertyName, $classMetadata->associationMappings);
+        $classMetadata->fieldMappings = $this->removeMappingByKey($propertyName, $classMetadata->fieldMappings);
+        $classMetadata->columnNames = $this->removeMappingByKey($propertyName, $classMetadata->columnNames);
+        $classMetadata->fieldNames = $this->removeMappingByValue($propertyName, $classMetadata->fieldNames);
+    }
+
+    /**
+     * @param string $propertyName
+     * @param array $mapping
+     * @return array
+     */
+    protected function removeMappingByKey(string $propertyName, array $mapping): array
+    {
+        if (array_key_exists($propertyName, $mapping)) {
+            unset($mapping[$propertyName]);
+        }
+
+        return $mapping;
+    }
+
+    /**
+     * @param string $propertyName
+     * @param array $mapping
+     * @return array
+     */
+    protected function removeMappingByValue(string $propertyName, array $mapping): array
+    {
+        $key = array_search($propertyName, $mapping, true);
+
+        if ($key !== false) {
+            unset($mapping[$key]);
+        }
+
+        return $mapping;
+    }
+}
