+++ upgrade-notes/backend_20250326_125354.md
@@ -1,0 +1,3 @@
+#### add the ability to easily remove entity columns from parent classes ([#2368](https://github.com/shopsys/shopsys/pull/2368))
+
+- see #project-base-diff to update your project
